DRY up code extracting the Slack logic
